Move kudos_streak and time_range to service obj
